seperated the hackathon dropdown and the idea loading
