+++ hms-app/src/components/card-details/CategoryDetails.tsx
