+++ hms-app/src/components/card-details/HackathonDetails.tsx
