+++ hms-app/src/components/card-details/IdeaDetails.tsx
