+++ hms-app/src/components/card-details/ParticipantDetails.tsx
