+++ hms-app/src/components/lists/AllHackathonList.tsx
@@ -2,7 +2,7 @@ import { Group, Button, createStyles, Card, Accordion } from '@mantine/core'
 import React, { useEffect, useState } from 'react'
 import { getListOfHackathons } from '../../actions/HackathonActions'
 import { HackathonPreview } from '../../common/types'
-import HackathonDetails from '../card- details/HackathonDetails'
+import HackathonDetails from '../card-details/HackathonDetails'
 
 const useStyles = createStyles((theme) => ({
   card: {
